Move LanguageServer alias to correct package.
